+++ src/Ext/Techno/Body.Update.cpp
@@ -5,7 +5,9 @@
 
 #include <Ext/Anim/Body.h>
 #include <Ext/Bullet/Body.h>
+#include <Ext/WarheadType/Body.h>
 #include <Ext/House/Body.h>
+#include <Ext/WeaponType/Body.h>
 #include <Utilities/EnumFunctions.h>
 #include <Utilities/AresFunctions.h>
 
@@ -688,4 +690,209 @@ void TechnoExt::UpdateSharedAmmo(TechnoClass* pThis)
 			}
 		}
 	}
+}
+
+void TechnoExt::ExtData::UpdateDelayFire()
+{
+	if (!this)
+		return;
+
+	auto const pThis = this->OwnerObject();
+	if (!pThis)
+		return;
+
+	if (!this->DelayedFire_Charging)
+		return;
+
+	auto clearVariables = [this]()
+	{
+		this->DelayedFire_Duration = -1;
+		this->DelayedFire_WeaponIndex = -1;
+		this->DelayedFire_DurationTimer.Stop();
+
+		if (this->DelayedFire_Anim)
+		{
+			if (this->DelayedFire_Anim->Type && !this->DelayedFire_Anim->InLimbo) // This anim doesn't have type pointer, just detach it
+			{
+				this->DelayedFire_Anim->TimeToDie = true;
+				this->DelayedFire_Anim->Limbo();
+			}
+
+			this->DelayedFire_Anim = nullptr;
+		}
+	};
+
+	// Disable the logic if the object isn't attacking
+	if (!pThis->Target || pThis->GetCurrentMission() != Mission::Attack)
+	{
+		clearVariables();
+		this->DelayedFire_Charged = false;
+		this->DelayedFire_Charging = false;
+
+		return;
+	}
+
+	int weaponIndex = pThis->SelectWeapon(pThis->Target);
+
+	// Check if weapon was changed and reset values if needed
+	if (this->DelayedFire_WeaponIndex != weaponIndex)
+	{
+		clearVariables();
+		return;
+	}
+
+	auto pWeaponType = pThis->GetWeapon(weaponIndex)->WeaponType;
+	if (!pWeaponType)
+		return;
+
+	auto pWeaponTypeExt = WeaponTypeExt::ExtMap.Find(pWeaponType);
+	if (!pWeaponTypeExt)
+		return;
+
+	if (this->DelayedFire_Charged)
+		return;
+
+	if (!pWeaponTypeExt->DelayedFire_Anim.isset())
+	{
+		clearVariables();
+		return;
+	}
+
+	CoordStruct animLocation = pThis->Location;
+
+	if (pWeaponTypeExt->DelayedFire_Anim_UseFLH)
+		animLocation = TechnoExt::GetFLHAbsoluteCoords(pThis, pThis->GetTechnoType()->Weapon[this->DelayedFire_WeaponIndex].FLH, pThis->HasTurret());
+
+	auto currentActiveAnim = this->DelayedFire_Anim;
+	int delayedFire_Duration = pWeaponTypeExt->DelayedFire_Duration.Get() > 0 ? pWeaponTypeExt->DelayedFire_Duration.Get() : 0;
+
+	if (this->DelayedFire_WeaponIndex >= 0)
+	{
+		// Check if weapon was changed and reset values
+		if (this->DelayedFire_WeaponIndex != weaponIndex)
+		{
+			clearVariables();
+			this->DelayedFire_Charging = false;
+			this->DelayedFire_Charged = false;
+
+			return;
+		}
+
+		// Check animation status
+		if (currentActiveAnim)
+		{
+			// If animation finished end these checks
+			if (currentActiveAnim->InLimbo)
+			{
+				// Disabling this logic, so the weapon can be fired
+				this->DelayedFire_Charging = false;
+				this->DelayedFire_Charged = true;
+				clearVariables();
+
+				if (pWeaponTypeExt->DelayedFire_PostAnim.isset())
+				{
+					if (auto pAnim = GameCreate<AnimClass>(pWeaponTypeExt->DelayedFire_PostAnim.Get(), animLocation))
+					{
+						this->DelayedFire_PostAnim = pAnim;
+						this->DelayedFire_PostAnim->SetOwnerObject(pThis);
+					}
+				}
+			}
+		}
+	}
+
+	if (currentActiveAnim)
+	{
+		// If is set a duration limit then this countdown takes preference over the original animation duration
+		if (delayedFire_Duration > 0)
+		{
+			if (!this->DelayedFire_DurationTimer.IsTicking())
+			{
+				// Execute only if timer has been stopped or not started
+				this->DelayedFire_DurationTimer.Start(delayedFire_Duration);
+				this->DelayedFire_Charging = true;
+				this->DelayedFire_Charged = false;
+				return;
+			}
+			else if (this->DelayedFire_DurationTimer.Completed())
+			{
+				// Execute only if timer has ran out after being started
+				// Delayed suicide
+				if (pWeaponTypeExt->DelayedFire_Suicide.Get())
+				{
+					pThis->ReceiveDamage(&pThis->Health, 0, RulesClass::Instance->C4Warhead, nullptr, true, false, nullptr);
+					WeaponTypeExt::DetonateAt(pWeaponType, pThis, pThis);
+				}
+
+				// Play an optional animation after the end of the main animation
+				if (pWeaponTypeExt->DelayedFire_PostAnim.isset())
+				{
+					if (auto pAnim = GameCreate<AnimClass>(pWeaponTypeExt->DelayedFire_PostAnim.Get(), animLocation))
+					{
+						this->DelayedFire_PostAnim = pAnim;
+						this->DelayedFire_PostAnim->SetOwnerObject(pThis);
+					}
+				}
+
+				// Disabling this logic, so the weapon can be fired
+				this->DelayedFire_Charging = false;
+				this->DelayedFire_Charged = true;
+				clearVariables();
+
+				return;
+			}
+			else
+			{
+				// Countdown still runs, can't fire until completion
+				return;
+			}
+		}
+		else
+		{
+			this->DelayedFire_DurationTimer.Stop();
+		}
+
+		if (this->DelayedFire_Charged)
+		{
+			// Disabling this logic, so the weapon can be fired
+			clearVariables();
+			this->DelayedFire_Charging = false;
+			this->DelayedFire_Charged = true;
+
+			// Delayed suicide
+			if (pWeaponTypeExt->DelayedFire_Suicide.Get())
+			{
+				pThis->ReceiveDamage(&pThis->Health, 0, RulesClass::Instance->C4Warhead, nullptr, true, false, nullptr);
+				WeaponTypeExt::DetonateAt(pWeaponType, pThis, pThis);
+			}
+			return;
+		}
+
+		return;
+	}
+
+	// Setting DelayFire from zero
+	clearVariables();
+	const auto pAnimType = pWeaponTypeExt->DelayedFire_Anim.Get();
+
+	auto pAnim = GameCreate<AnimClass>(pAnimType, animLocation);
+	if (!pAnim)
+	{
+		Debug::Log("DelayedFire animation [%s] for [%s] can't be created. Logic disabled.\n", pAnimType->ID, pThis->GetTechnoType()->ID);
+		clearVariables();
+		this->DelayedFire_Charging = false;
+		this->DelayedFire_Charged = true;
+
+		return;
+	}
+
+	this->DelayedFire_Anim = pAnim;
+	this->DelayedFire_Anim->SetOwnerObject(pThis);
+	this->DelayedFire_WeaponIndex = weaponIndex;
+	this->DelayedFire_Duration = delayedFire_Duration;
+	this->DelayedFire_Charging = true;
+	this->DelayedFire_Charged = false;
+
+	if (delayedFire_Duration > 0)
+		this->DelayedFire_DurationTimer.Start(delayedFire_Duration);
 }