+++ src/Ext/Techno/Body.cpp
@@ -659,6 +659,13 @@ void TechnoExt::ExtData::Serialize(T& Stm)
 		.Process(this->DeployFireTimer)
 		.Process(this->ForceFullRearmDelay)
 		.Process(this->WHAnimRemainingCreationInterval)
+		.Process(this->DelayedFire_Charging)
+		.Process(this->DelayedFire_Charged)
+		.Process(this->DelayedFire_Anim)
+		.Process(this->DelayedFire_PostAnim)
+		.Process(this->DelayedFire_Duration)
+		.Process(this->DelayedFire_WeaponIndex)
+		.Process(this->DelayedFire_DurationTimer)
 		;
 }