+++ src/Ext/Techno/Hooks.cpp
@@ -33,6 +33,7 @@ DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
 	pExt->DepletedAmmoActions();
 
 	TechnoExt::ApplyMindControlRangeLimit(pThis);
+	pExt->UpdateDelayFire();
 
 	return 0;
 }